requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pandas>=2.2.0
pdfplumber>=0.11.0
PyPDF2>=3.0.1
python-docx>=1.1.0
spacy>=3.7.4
werkzeug>=3.0.1
plotly>=5.18.0
